webapi/tct-extra-html5-tests/rendering/bindings/the-keygen-element-0/htmlkeygenelement_attributes.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xu, YuhanX <yuhanx.xu@intel.com>

-->
<html>
  <head>
    <title>keygen Test: htmlkeygenelement_attributes</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/forms.html#the-keygen-element">
    <meta name="flags" content="">
    <meta name="assert" content="Check if HTMLKeygenElement attributes">
    <script src="../../../resources/testharness.js"></script>
    <script src="../../../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id ="log"></div>
    <keygen id="security" />
    <script type="text/javascript">
        var security = document.getElementById("security");
        [
            ["autofocus", "boolean"],
            ["challenge", "DOMString"],
            ["disabled", "boolean"],
            ["form", "object"],
            ["keytype", "DOMString"],
            ["name", "DOMString"],
            ["type", "DOMString"],
            ["willValidate", "boolean"],
            ["validity", "object"],
            ["validationMessage", "DOMString"],
            ["checkValidity", "function"],
            ["setCustomValidity", "function"],
            ["labels", "object"],

        ].forEach(function(attr) {
            var name = attr[0];
            var type = attr[1];

            // existence check
            test(function() {
                assert_true(name in security, name + " attribute in audio element");
            }, "Check if HTMLKeygenElement." + name + " exists");

            switch(type) {
            case "boolean":
                test(function() {
                    assert_equals(typeof security[name], "boolean", name + " attribute of type");
                }, "Check if HTMLKeygenElement." + name + " is of type boolean");
                break;

            case "DOMString":
                test(function() {
                    assert_equals(typeof security[name], "string", name + " attribute of type");
                }, "Check if HTMLKeygenElement." + name + " is of type string");
                break;

            case "function":
                test(function() {
                    assert_equals(typeof security[name], "function", name + " attribute of type");
                }, "Check if HTMLKeygenElement." + name + " is of type function");
                break;

            case "object":
                test(function() {
                    assert_equals(typeof security[name], "object", name + " attribute of type");
                }, "Check if HTMLKeygenElement." + name + " is of type object");
                break;

            default:
                break;
            }

        });
    </script>
  </body>
</html>
